+++ flutter/lib/src/event_processor/native_app_start_event_processor.dart
@@ -1,16 +1,24 @@
+// ignore_for_file: invalid_use_of_internal_member
+
 import 'dart:async';
 
-import 'package:sentry/sentry.dart';
-
+import '../../sentry_flutter.dart';
 import '../integrations/integrations.dart';
 import '../native/sentry_native.dart';
+
+// ignore: implementation_imports
+import 'package:sentry/src/sentry_tracer.dart';
 
 /// EventProcessor that enriches [SentryTransaction] objects with app start
 /// measurement.
 class NativeAppStartEventProcessor implements EventProcessor {
   final SentryNative _native;
+  final Hub _hub;
 
-  NativeAppStartEventProcessor(this._native);
+  NativeAppStartEventProcessor(
+    this._native, {
+    Hub? hub,
+  }) : _hub = hub ?? HubAdapter();
 
   @override
   Future<SentryEvent?> apply(SentryEvent event, Hint hint) async {
@@ -25,6 +33,82 @@ class NativeAppStartEventProcessor implements EventProcessor {
       event.measurements[measurement.name] = measurement;
       _native.didAddAppStartMeasurement = true;
     }
+
+    if (appStartInfo != null) {
+      await _attachAppStartSpans(appStartInfo, event.tracer);
+    }
+
     return event;
+  }
+
+  Future<void> _attachAppStartSpans(
+      AppStartInfo appStartInfo, SentryTracer transaction) async {
+    final transactionTraceId = transaction.context.traceId;
+
+    final appStartSpan = await _createAndFinishSpan(
+        tracer: transaction,
+        operation: appStartInfo.appStartTypeOperation,
+        description: appStartInfo.appStartTypeDescription,
+        parentSpanId: transaction.context.spanId,
+        traceId: transactionTraceId,
+        startTimestamp: appStartInfo.start,
+        endTimestamp: appStartInfo.end);
+
+    final pluginRegistrationSpan = await _createAndFinishSpan(
+        tracer: transaction,
+        operation: appStartInfo.appStartTypeOperation,
+        description: appStartInfo.pluginRegistrationDescription,
+        parentSpanId: appStartSpan.context.spanId,
+        traceId: transactionTraceId,
+        startTimestamp: appStartInfo.start,
+        endTimestamp: appStartInfo.pluginRegistration);
+
+    final mainIsolateSetupSpan = await _createAndFinishSpan(
+        tracer: transaction,
+        operation: appStartInfo.appStartTypeOperation,
+        description: appStartInfo.mainIsolateSetupDescription,
+        parentSpanId: appStartSpan.context.spanId,
+        traceId: transactionTraceId,
+        startTimestamp: appStartInfo.pluginRegistration,
+        endTimestamp: appStartInfo.mainIsolateStart);
+
+    final firstFrameRenderSpan = await _createAndFinishSpan(
+        tracer: transaction,
+        operation: appStartInfo.appStartTypeOperation,
+        description: appStartInfo.firstFrameRenderDescription,
+        parentSpanId: appStartSpan.context.spanId,
+        traceId: transactionTraceId,
+        startTimestamp: appStartInfo.mainIsolateStart,
+        endTimestamp: appStartInfo.end);
+
+    transaction.children.addAll([
+      appStartSpan,
+      pluginRegistrationSpan,
+      mainIsolateSetupSpan,
+      firstFrameRenderSpan
+    ]);
+  }
+
+  Future<SentrySpan> _createAndFinishSpan({
+    required SentryTracer tracer,
+    required String operation,
+    required String description,
+    required SpanId parentSpanId,
+    required SentryId traceId,
+    required DateTime startTimestamp,
+    required DateTime endTimestamp,
+  }) async {
+    final span = SentrySpan(
+        tracer,
+        SentrySpanContext(
+          operation: operation,
+          description: description,
+          parentSpanId: parentSpanId,
+          traceId: traceId,
+        ),
+        _hub,
+        startTimestamp: startTimestamp);
+    await span.finish(endTimestamp: endTimestamp);
+    return span;
   }
 }