+++ flutter/lib/src/integrations/native_app_start_integration.dart
@@ -53,7 +53,11 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
     if (isIntegrationTest) {
       final appStartInfo = AppStartInfo(AppStartType.cold,
           start: DateTime.now(),
-          end: DateTime.now().add(const Duration(milliseconds: 100)));
+          end: DateTime.now().add(const Duration(milliseconds: 100)),
+          pluginRegistration:
+              DateTime.now().add(const Duration(milliseconds: 50)),
+          mainIsolateStart:
+              DateTime.now().add(const Duration(milliseconds: 60)));
       setAppStartInfo(appStartInfo);
       return;
     }
@@ -67,16 +71,22 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
         // We only assign the current time if it's not already set - this is useful in tests
         // ignore: invalid_use_of_internal_member
         _native.appStartEnd ??= options.clock();
-        final appStartEnd = _native.appStartEnd;
+        final appStartEndDateTime = _native.appStartEnd;
         final nativeAppStart = await _native.fetchNativeAppStart();
+        final pluginRegistrationTime = nativeAppStart?.pluginRegistrationTime;
+        final mainIsolateStartDateTime = SentryFlutter.mainIsolateStartTime;
 
-        if (nativeAppStart == null || appStartEnd == null) {
+        if (nativeAppStart == null ||
+            appStartEndDateTime == null ||
+            pluginRegistrationTime == null) {
           return;
         }
 
         final appStartDateTime = DateTime.fromMillisecondsSinceEpoch(
             nativeAppStart.appStartTime.toInt());
-        final duration = appStartEnd.difference(appStartDateTime);
+        final duration = appStartEndDateTime.difference(appStartDateTime);
+        final pluginRegistrationDateTime =
+            DateTime.fromMillisecondsSinceEpoch(pluginRegistrationTime);
 
         // We filter out app start more than 60s.
         // This could be due to many different reasons.
@@ -93,9 +103,11 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
 
         final appStartInfo = AppStartInfo(
             nativeAppStart.isColdStart ? AppStartType.cold : AppStartType.warm,
-            start: DateTime.fromMillisecondsSinceEpoch(
-                nativeAppStart.appStartTime.toInt()),
-            end: appStartEnd);
+            start: appStartDateTime,
+            end: appStartEndDateTime,
+            pluginRegistration: pluginRegistrationDateTime,
+            mainIsolateStart: mainIsolateStartDateTime);
+
         setAppStartInfo(appStartInfo);
       });
     }
@@ -109,11 +121,18 @@ class NativeAppStartIntegration extends Integration<SentryFlutterOptions> {
 enum AppStartType { cold, warm }
 
 class AppStartInfo {
-  AppStartInfo(this.type, {required this.start, required this.end});
+  AppStartInfo(this.type,
+      {required this.start,
+      required this.end,
+      required this.pluginRegistration,
+      required this.mainIsolateStart});
 
   final AppStartType type;
   final DateTime start;
   final DateTime end;
+  final DateTime pluginRegistration;
+  final DateTime mainIsolateStart;
+
   Duration get duration => end.difference(start);
 
   SentryMeasurement toMeasurement() {
@@ -121,4 +140,12 @@ class AppStartInfo {
         ? SentryMeasurement.coldAppStart(duration)
         : SentryMeasurement.warmAppStart(duration);
   }
+
+  String get appStartTypeOperation => 'app.start.${type.name}';
+
+  String get appStartTypeDescription =>
+      type == AppStartType.cold ? 'Cold start' : 'Warm start';
+  final pluginRegistrationDescription = 'App start to plugin registration';
+  final mainIsolateSetupDescription = 'Main isolate setup';
+  final firstFrameRenderDescription = 'First frame render';
 }