+++ flutter/lib/src/sentry_flutter.dart
@@ -34,6 +34,11 @@ typedef FlutterOptionsConfiguration = FutureOr<void> Function(
 mixin SentryFlutter {
   static const _channel = MethodChannel('sentry_flutter');
 
+  /// Represents the time when the main isolate is set up and ready to execute.
+  @internal
+  // ignore: invalid_use_of_internal_member
+  static DateTime mainIsolateStartTime = getUtcDateTime();
+
   static Future<void> init(
     FlutterOptionsConfiguration optionsConfiguration, {
     AppRunner? appRunner,