+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Adds app start spans to first transaction ([#2009](https://github.com/getsentry/sentry-dart/pull/2009))
+
 ## 8.1.0
 
-### Feature
+### Features
 
 - Set snapshot to `true` if stacktrace is not provided ([#2000](https://github.com/getsentry/sentry-dart/pull/2000))
   - If the stacktrace is not provided, the Sentry SDK will fetch the current stacktrace via `StackTrace.current` and the snapshot will be set to `true` - **this may change the grouping behavior**